1.2
完成入库单的UI及数据控制

+++ Control/meteData.py
@@ -7,20 +7,27 @@ Created on 2015年8月5日
 
 
 import sqlite3
+from PyQt4.QtGui import QMessageBox
 
 
 def meteData():
-    dataConnect = sqlite3.connect('data.db')
-    cursor = dataConnect.cursor()
-    cursor.execute(u'DROP TABLE IF EXISTS rkd')
-    cursor.execute(u'create table rkd (ID integer PRIMARY KEY autoincrement,DocNO c,Custom c,Item c,Weight e,InDate i,Stock f  ) ')
-    cursor.execute(u'pragma table_info(rkd)')
+    answer=QMessageBox.question(None,u"数据库初始化",  
+                                u"已到达文档结尾,是否从头查找?",  
+                                QMessageBox.Ok|QMessageBox.Cancel,  
+                                QMessageBox.Ok)
     
-    print cursor.fetchall()
-    cursor.close()
-    dataConnect.close()
-    print u'数据库初始化成功'
-    
+    if answer == QMessageBox.Ok:
+        dataConnect = sqlite3.connect('data.db')
+        cursor = dataConnect.cursor()
+        cursor.execute(u'DROP TABLE IF EXISTS rkd')
+        cursor.execute(u'create table rkd (ID integer PRIMARY KEY autoincrement,DocNO c,Custom c,Item c,Weight e,InDate i,Stock f  ) ')
+        cursor.execute(u'pragma table_info(rkd)')
+        cursor.fetchall()
+        cursor.close()
+        dataConnect.close()
+        QMessageBox.information(None,u'数据库初始化',u'初始化数据库成功！')
+    else:
+        return
